refactoring moved videoMetadata extraction to seperate service
