+++ src/utils/upload/index.ts
@@ -49,3 +49,15 @@ export function segregateMetadataAndMediaFiles(
     });
     return { mediaFiles, metadataJSONFiles };
 }
+
+export function getUNIXTime(dateTime: Date) {
+    if (!dateTime || isNaN(dateTime.getTime())) {
+        return null;
+    }
+    const unixTime = dateTime.getTime() * 1000;
+    if (unixTime <= 0) {
+        return null;
+    } else {
+        return unixTime;
+    }
+}
